split up files and remove unecessary dependencies
