+++ src/interface/mesh_data.hpp
@@ -29,6 +29,7 @@
 #include "mesh/meshblock.hpp"
 #include "mesh/meshblock_pack.hpp"
 #include "pack/sparse_pack_base.hpp"
+#include "pack/sparse_pack_cache.hpp"
 #include "pack/swarm_pack_base.hpp"
 #include "utils/communication_buffer.hpp"
 #include "utils/error_checking.hpp"
