+++ src/pack/pack_descriptor.cpp
@@ -1,0 +1,34 @@
+//========================================================================================
+// (C) (or copyright) 2020-2025. Triad National Security, LLC. All rights reserved.
+//
+// This program was produced under U.S. Government contract 89233218CNA000001 for Los
+// Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
+// for the U.S. Department of Energy/National Nuclear Security Administration. All rights
+// in the program are reserved by Triad National Security, LLC, and the U.S. Department
+// of Energy/National Nuclear Security Administration. The Government is granted for
+// itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide
+// license in this material to reproduce, prepare derivative works, distribute copies to
+// the public, perform publicly and display publicly, and to permit others to do so.
+//========================================================================================
+
+#include <cstdio>
+#include <string>
+
+#include "pack/pack_descriptor.hpp"
+
+namespace parthenon {
+namespace impl {
+
+void PackDescriptor::Print() const {
+  printf("--------------------\n");
+  for (int i = 0; i < var_group_names.size(); ++i) {
+    printf("group name: %s\n", var_group_names[i].c_str());
+    printf("--------------------\n");
+    for (const auto &[var_name, uid] : var_groups[i]) {
+      printf("%s\n", var_name.label().c_str());
+    }
+  }
+  printf("--------------------\n");
+}
+} // namespace impl
+} // namespace parthenon
