+++ src/pack/sparse_pack_base.cpp
@@ -31,27 +31,25 @@
 #include "kokkos_abstraction.hpp"
 #include "pack/pack_utils.hpp"
 #include "pack/sparse_pack_base.hpp"
+#include "pack/sparse_pack_cache.hpp"
 #include "utils/utils.hpp"
-namespace parthenon {
-namespace impl {
-
-void PackDescriptor::Print() const {
-  printf("--------------------\n");
-  for (int i = 0; i < var_group_names.size(); ++i) {
-    printf("group name: %s\n", var_group_names[i].c_str());
-    printf("--------------------\n");
-    for (const auto &[var_name, uid] : var_groups[i]) {
-      printf("%s\n", var_name.label().c_str());
-    }
-  }
-  printf("--------------------\n");
-}
-} // namespace impl
-} // namespace parthenon
 
 namespace parthenon {
 
 using namespace impl;
+
+template <class T>
+SparsePackBase SparsePackBase::GetPack(T *pmd, const impl::PackDescriptor &desc,
+                                              const std::vector<bool> &include_block) {
+  auto &cache = pmd->GetSparsePackCache();
+  return cache.Get(pmd, desc, include_block);
+}
+template SparsePackBase SparsePackBase::GetPack<MeshBlockData<Real>>(MeshBlockData<Real>*, 
+                                                                     const impl::PackDescriptor&,
+                                                                     const std::vector<bool>&);
+template SparsePackBase SparsePackBase::GetPack<MeshData<Real>>(MeshData<Real>*, 
+                                                                const impl::PackDescriptor&,
+                                                                const std::vector<bool>&);
 
 SparsePackIdxMap SparsePackBase::GetIdxMap(const impl::PackDescriptor &desc) {
   SparsePackIdxMap map;
@@ -350,52 +348,5 @@ SparsePackBase::Build<MeshBlockData<Real>>(MeshBlockData<Real> *, const PackDesc
 template SparsePackBase SparsePackBase::Build<MeshData<Real>>(MeshData<Real> *,
                                                               const PackDescriptor &,
                                                               const std::vector<bool> &);
-
-template <class T>
-SparsePackBase &SparsePackCache::Get(T *pmd, const PackDescriptor &desc,
-                                     const std::vector<bool> &include_block) {
-  if (pack_map.count(desc.identifier) > 0) {
-    auto &desc_pack_map = pack_map[desc.identifier];
-    if (desc_pack_map.count(include_block)) {
-      auto &cache_tuple = desc_pack_map[include_block];
-      auto &pack = std::get<0>(cache_tuple);
-      auto alloc_status_in = SparsePackBase::GetAllocStatus(pmd, desc, include_block);
-      auto &alloc_status = std::get<1>(cache_tuple);
-      if (alloc_status.size() != alloc_status_in.size())
-        return BuildAndAdd(pmd, desc, include_block);
-      for (int i = 0; i < alloc_status_in.size(); ++i) {
-        if (alloc_status[i] != alloc_status_in[i])
-          return BuildAndAdd(pmd, desc, include_block);
-      }
-      // Cached version is not stale, so just return a reference to it
-      return std::get<0>(cache_tuple);
-    }
-  }
-  return BuildAndAdd(pmd, desc, include_block);
-}
-template SparsePackBase &SparsePackCache::Get<MeshData<Real>>(MeshData<Real> *,
-                                                              const PackDescriptor &,
-                                                              const std::vector<bool> &);
-template SparsePackBase &
-SparsePackCache::Get<MeshBlockData<Real>>(MeshBlockData<Real> *, const PackDescriptor &,
-                                          const std::vector<bool> &);
-
-template <class T>
-SparsePackBase &SparsePackCache::BuildAndAdd(T *pmd, const PackDescriptor &desc,
-                                             const std::vector<bool> &include_block) {
-  if (pack_map.count(desc.identifier) == 0) pack_map[desc.identifier] = desc_pack_map_t();
-
-  desc_pack_map_t &desc_pack_map = pack_map[desc.identifier];
-  desc_pack_map[include_block] = {
-      SparsePackBase::Build(pmd, desc, include_block),
-      SparsePackBase::GetAllocStatus(pmd, desc, include_block)};
-
-  return std::get<0>(desc_pack_map[include_block]);
-}
-template SparsePackBase &
-SparsePackCache::BuildAndAdd<MeshData<Real>>(MeshData<Real> *, const PackDescriptor &,
-                                             const std::vector<bool> &);
-template SparsePackBase &SparsePackCache::BuildAndAdd<MeshBlockData<Real>>(
-    MeshBlockData<Real> *, const PackDescriptor &, const std::vector<bool> &);
 
 } // namespace parthenon
