+++ src/pack/sparse_pack_base.hpp
@@ -13,38 +13,17 @@
 #ifndef PACK_SPARSE_PACK_BASE_HPP_
 #define PACK_SPARSE_PACK_BASE_HPP_
 
-#include <algorithm>
-#include <functional>
-#include <limits>
-#include <map>
-#include <memory>
-#include <set>
-#include <string>
-#include <tuple>
-#include <type_traits>
-#include <unordered_map>
-#include <utility>
 #include <vector>
 
 #include "coordinates/coordinates.hpp"
-#include "interface/state_descriptor.hpp"
 #include "interface/variable.hpp"
 #include "interface/variable_state.hpp"
 #include "kokkos_abstraction.hpp"
-#include "utils/utils.hpp"
+#include "pack/pack_descriptor.hpp"
 
 namespace parthenon {
-class SparsePackCache;
-namespace impl {
-class PackDescriptor;
-}
-
 // Map for going from variable names to sparse pack variable indices
 using SparsePackIdxMap = std::unordered_map<std::string, std::size_t>;
-
-class StateDescriptor;
-
-enum class PDOpt { WithFluxes, Coarse, Flatten };
 
 class SparsePackBase {
  public:
@@ -70,10 +49,7 @@ class SparsePackBase {
   // from the cache in pmd. The cache itself handles the all of this logic
   template <class T>
   static SparsePackBase GetPack(T *pmd, const impl::PackDescriptor &desc,
-                                const std::vector<bool> &include_block) {
-    auto &cache = pmd->GetSparsePackCache();
-    return cache.Get(pmd, desc, include_block);
-  }
+                                const std::vector<bool> &include_block); 
 
   // Return a map from variable names to pack variable indices
   static SparsePackIdxMap GetIdxMap(const impl::PackDescriptor &desc);
@@ -107,147 +83,6 @@ class SparsePackBase {
   int nvar_;
   int size_;
 };
-
-// Object for cacheing sparse packs in MeshData and MeshBlockData objects. This
-// handles checking for a pre-existing pack and creating a new SparsePackBase if
-// a cached pack is unavailable. Essentially, this operates as a map from
-// `PackDescriptor` to `SparsePackBase`
-class SparsePackCache {
- public:
-  std::size_t size() const {
-    std::size_t size{0};
-    for (const auto &[key, desc_pack_map] : pack_map) {
-      size += desc_pack_map.size();
-    }
-    return size;
-  }
-
-  void clear() { pack_map.clear(); }
-
- protected:
-  template <class T>
-  SparsePackBase &Get(T *pmd, const impl::PackDescriptor &desc,
-                      const std::vector<bool> &include_block);
-
-  template <class T>
-  SparsePackBase &BuildAndAdd(T *pmd, const impl::PackDescriptor &desc,
-                              const std::vector<bool> &include_block);
-  // For a given pack descriptor, this contains a map from a vector containing the
-  // requested included blocks to a cached pack, along with the allocation status of
-  // varirables in the pack
-  using desc_pack_map_t =
-      std::map<std::vector<bool>, std::tuple<SparsePackBase, SparsePackBase::alloc_t>>;
-  // Map from pack descriptor id to map from included blocks
-  using pack_map_t = std::unordered_map<std::string, desc_pack_map_t>;
-  pack_map_t pack_map;
-
-  friend class SparsePackBase;
-};
-
-namespace impl {
-struct PackDescriptor {
-  using VariableGroup_t = std::vector<std::pair<VarID, Uid_t>>;
-  using SelectorFunction_t = std::function<bool(int, const VarID &, const Metadata &)>;
-  using SelectorFunctionUid_t = std::function<bool(int, const Uid_t &, const Metadata &)>;
-
-  void Print() const;
-
-  // default constructor needed for certain use cases
-  PackDescriptor()
-      : nvar_groups(0), var_group_names({}), var_groups({}), with_fluxes(false),
-        coarse(false), flat(false), identifier(""), nvar_tot(0) {}
-
-  template <class GROUP_t, class SELECTOR_t>
-  PackDescriptor(StateDescriptor *psd, const std::vector<GROUP_t> &var_groups_in,
-                 const SELECTOR_t &selector, const std::set<PDOpt> &options)
-      : nvar_groups(var_groups_in.size()), var_group_names(MakeGroupNames(var_groups_in)),
-        var_groups(BuildUids(var_groups_in.size(), psd, selector)),
-        with_fluxes(options.count(PDOpt::WithFluxes)),
-        coarse(options.count(PDOpt::Coarse)), flat(options.count(PDOpt::Flatten)),
-        identifier(GetIdentifier()), nvar_tot(GetNVarsTotal(var_groups)) {
-    PARTHENON_REQUIRE(!(with_fluxes && coarse),
-                      "Probably shouldn't be making a coarse pack with fine fluxes.");
-  }
-
-  const int nvar_groups;
-  const std::vector<std::string> var_group_names;
-  const std::vector<VariableGroup_t> var_groups;
-  const bool with_fluxes;
-  const bool coarse;
-  const bool flat;
-  const std::string identifier;
-  const std::size_t nvar_tot;
-
- private:
-  static int GetNVarsTotal(const std::vector<VariableGroup_t> &var_groups) {
-    int nvar_tot = 0;
-    for (const auto &group : var_groups) {
-      for (const auto &[a, b] : group) {
-        nvar_tot++;
-      }
-    }
-    return nvar_tot;
-  }
-  std::string GetIdentifier() {
-    std::string ident("");
-    for (const auto &vgroup : var_groups) {
-      for (const auto &[vid, uid] : vgroup) {
-        ident += std::to_string(uid) + "_";
-      }
-      ident += "|";
-    }
-    ident += std::to_string(with_fluxes);
-    ident += std::to_string(coarse);
-    ident += std::to_string(flat);
-    return ident;
-  }
-  template <class FUNC_t>
-  std::vector<PackDescriptor::VariableGroup_t>
-  BuildUids(int nvgs, const StateDescriptor *const psd, const FUNC_t &selector) {
-    auto fields = psd->AllFields();
-    std::vector<VariableGroup_t> vgs(nvgs);
-    for (auto [id, md] : fields) {
-      for (int i = 0; i < nvgs; ++i) {
-        auto uid = Variable<Real>::GetUniqueID(id.label());
-        if constexpr (std::is_invocable<FUNC_t, int, VarID, Metadata>::value) {
-          if (selector(i, id, md)) {
-            vgs[i].push_back({id, uid});
-          }
-        } else if constexpr (std::is_invocable<FUNC_t, int, Uid_t, Metadata>::value) {
-          if (selector(i, uid, md)) {
-            vgs[i].push_back({id, uid});
-          }
-        } else {
-          PARTHENON_FAIL("Passing the wrong sort of selector.");
-        }
-      }
-    }
-    // Ensure ordering in terms of value of sparse indices
-    for (auto &vg : vgs) {
-      std::sort(vg.begin(), vg.end(), [](const auto &a, const auto &b) {
-        if (a.first.base_name == b.first.base_name)
-          return a.first.sparse_id < b.first.sparse_id;
-        return a.first.base_name < b.first.base_name;
-      });
-    }
-    return vgs;
-  }
-
-  template <class base_t>
-  std::vector<std::string> MakeGroupNames(const std::vector<base_t> &var_groups) {
-    if constexpr (std::is_same<base_t, std::string>::value) {
-      return var_groups;
-    } else if constexpr (std::is_same<base_t, Uid_t>::value) {
-      std::vector<std::string> var_group_names;
-      for (auto &vg : var_groups)
-        var_group_names.push_back(std::to_string(vg));
-      return var_group_names;
-    }
-    // silence compiler warnings about no return statement
-    return std::vector<std::string>();
-  }
-};
-} // namespace impl
 
 } // namespace parthenon
